FIX: arm type cast issue

+++ src/common_doca.c
@@ -213,7 +213,7 @@ void print_buffer_hex(const void *buffer, size_t length)
 
 doca_error_t sock_send_buffer(const void *rdma_conn_descriptor, size_t descriptor_size, int sock_fd)
 {
-    if (sock_write(sock_fd, &descriptor_size, sizeof(uint32_t)) != sizeof(uint32_t))
+    if (sock_write(sock_fd, &descriptor_size, sizeof(size_t)) != sizeof(size_t))
     {
         log_error("Error, send descriptor size\n");
         goto error;
@@ -241,7 +241,7 @@ doca_error_t sock_recv_buffer(void *rdma_conn_descriptor, size_t *descriptor_siz
                               int sock_fd)
 {
 
-    if (sock_read(sock_fd, descriptor_size, sizeof(uint32_t)) != sizeof(uint32_t))
+    if (sock_read(sock_fd, descriptor_size, sizeof(size_t)) != sizeof(size_t))
     {
         log_error("Error, recv descriptor size\n");
         goto error;
